move battle engine and pokemon
