+++ src/services/battle.js
@@ -6,29 +6,17 @@
 /**
  * @file
  * @author Elvis Wei
- * @date 2023
- * @section Description
  *
  * battle.js Handles all battle interactions from the user at a base level down to creating the teams.
  */
 const { v4: uuidv4 } = require("uuid");
-const { getOrSetDefault, formatMoney } = require("../utils/utils");
+const { Battle } = require("../battle/engine/Battle");
+const { getOrSetDefault } = require("../utils/utils");
 const {
-  pokemonConfig,
-  types: pokemonTypes,
-} = require("../config/pokemonConfig");
-const {
-  targetTypes,
-  targetPatterns,
-  targetPositions,
   statusConditions,
   moveTiers,
   calculateDamage,
-  typeAdvantages,
-  weatherConditions,
-  damageTypes,
 } = require("../config/battleConfig");
-const { battleEventEnum } = require("../enums/battleEnums");
 const {
   buildBattleEmbed,
   buildPveListEmbed,
@@ -45,20 +33,10 @@ const {
   buildBattleInfoActionRow,
 } = require("../components/battleInfoActionRow");
 const { getTrainer, addExpAndMoney, updateTrainer } = require("./trainer");
-const {
-  addPokemonExpAndEVs,
-  getPokemon,
-  calculatePokemonStats,
-} = require("./pokemon");
+const { addPokemonExpAndEVs, getPokemon } = require("./pokemon");
 const { logger } = require("../log");
 const { buildNextTurnActionRow } = require("../components/battleNextTurnRow");
 const { deleteState } = require("./state");
-const {
-  calculateEffectiveSpeed,
-  calculateEffectiveAccuracy,
-  calculateEffectiveEvasion,
-  getMoveIds,
-} = require("../utils/pokemonUtils");
 const {
   npcConfig,
   difficultyConfig,
@@ -70,19 +48,12 @@ const { buildScrollActionRow } = require("../components/scrollActionRow");
 const { getState } = require("./state");
 const { eventNames } = require("../config/eventConfig");
 const { buildIdConfigSelectRow } = require("../components/idConfigSelectRow");
-const {
-  drawIterable,
-  drawUniform,
-  drawDiscrete,
-} = require("../utils/gachaUtils");
+const { drawIterable, drawUniform } = require("../utils/gachaUtils");
 const { generateRandomPokemon } = require("./gacha");
 const { validateParty } = require("./party");
 const { addRewards, getRewardsString } = require("../utils/trainerUtils");
 const { getIdFromTowerStage } = require("../utils/battleUtils");
 const { getMove } = require("../battle/data/moveRegistry");
-const { getEffect } = require("../battle/data/effectRegistry");
-const { getAbility } = require("../battle/data/abilityRegistry");
-const { BattleEventHandler } = require("../battle/engine/events");
 
 class NPC {
   constructor(
@@ -449,2532 +420,6 @@ class RaidNPC extends NPC {
       pokemon._id = pokemon._id || uuidv4();
       this.party.pokemons[pokemonData.position - 1] = pokemon;
     }
-  }
-}
-
-class Pokemon {
-  /* battle;
-  pokemonData;
-  speciesId;
-  speciesData;
-  id;
-  userId;
-  originalUserId;
-  teamName;
-  name;
-  level;
-  hp;
-  maxHp;
-  atk;
-  batk;
-  def;
-  bdef;
-  spa;
-  bspa;
-  spd;
-  bspd;
-  spe;
-  bspe;
-  acc;
-  eva;
-  type1;
-  type2;
-  // effectId => { duration, args }
-  effectIds;
-  // moveId => { cooldown, disabled }
-  moveIds;
-  // { statusId. tuns active }
-  status;
-  // { abilityId, args }
-  ability;
-  combatReadiness;
-  position;
-  isFainted;
-  targetable;
-  hittable;
-  incapacitated;
-  restricted; */
-
-  constructor(battle, trainer, pokemonData, teamName, position) {
-    this.battle = battle;
-    this.speciesId = pokemonData.speciesId;
-    this.speciesData = pokemonConfig[this.speciesId];
-    // if battle has a set level, scale pokemon to that level
-    if (battle.level) {
-      battle.minLevel = Math.min(battle.minLevel, pokemonData.level);
-      pokemonData.level = battle.level;
-    }
-    // if battle has an equipment level, set all pokemons equipments to that level
-    if (battle.equipmentLevel) {
-      for (const equipment of Object.values(pokemonData.equipments)) {
-        equipment.level = battle.equipmentLevel;
-      }
-    }
-    pokemonData = calculatePokemonStats(pokemonData, this.speciesData);
-    this.pokemonData = pokemonData;
-    this.id = pokemonData._id.toString();
-    this.userId = trainer.userId;
-    this.originalUserId = trainer.userId;
-    this.teamName = teamName;
-    this.name = pokemonData.name;
-    this.hp = pokemonData.remainingHp || pokemonData.stats[0];
-    [this.maxHp = 0] = this.pokemonData.stats;
-    this.level = pokemonData.level;
-    [
-      this.atk = 0,
-      this.batk = 0,
-      this.def,
-      this.bdef = 0,
-      this.spa,
-      this.bspa = 0,
-      this.spd,
-      this.bspd = 0,
-      this.spe,
-      this.bspe = 0,
-    ] = [
-      pokemonData.stats[1],
-      pokemonData.stats[1],
-      pokemonData.stats[2],
-      pokemonData.stats[2],
-      pokemonData.stats[3],
-      pokemonData.stats[3],
-      pokemonData.stats[4],
-      pokemonData.stats[4],
-      pokemonData.stats[5],
-      pokemonData.stats[5],
-    ];
-    this.acc = 100;
-    this.eva = 100;
-    [this.type1 = null, this.type2 = null] = this.speciesData.type;
-    // map effectId => effect data (duration, args)
-    this.effectIds = {};
-    // map moveId => move data (cooldown, disabled)
-    this.addMoves(pokemonData);
-    this.setAbility(pokemonData);
-    this.status = {
-      statusId: null,
-      turns: 0,
-    };
-    this.combatReadiness = 0;
-    this.position = position;
-    this.isFainted = false;
-    this.targetable = true;
-    this.hittable = true;
-    this.incapacitated = false;
-    this.restricted = false;
-  }
-
-  addMoves(pokemonData) {
-    this.moveIds = getMoveIds(pokemonData).reduce((acc, moveId) => {
-      acc[moveId] = {
-        cooldown: 0,
-        disabled: false,
-      };
-      return acc;
-    }, {});
-  }
-
-  setAbility(pokemonData) {
-    this.ability = {
-      abilityId: pokemonData.abilityId,
-    };
-  }
-
-  transformInto(speciesId, { abilityId = null, moveIds = [] } = {}) {
-    const oldName = this.name;
-    // remove ability
-    this.removeAbility();
-
-    // get old stat ratios (excpet hp)
-    const atkRatio = this.atk / this.batk;
-    const defRatio = this.def / this.bdef;
-    const spaRatio = this.spa / this.bspa;
-    const spdRatio = this.spd / this.bspd;
-    const speRatio = this.spe / this.bspe;
-
-    // set new species values
-    this.speciesId = speciesId;
-    this.speciesData = pokemonConfig[this.speciesId];
-    [this.type1, this.type2 = null] = this.speciesData.type;
-    this.type2 = this.speciesData.type[1] || null;
-    this.name = this.speciesData.name;
-
-    // set pokemon data object
-    this.pokemonData.speciesId = this.speciesId;
-    this.pokemonData.name = this.speciesData.name;
-    this.pokemonData.moveIds = moveIds;
-    this.pokemonData.abilityId =
-      abilityId || drawDiscrete(this.speciesData.abilities, 1)[0];
-    calculatePokemonStats(this.pokemonData, this.speciesData);
-
-    // set stats (except hp)
-    [, this.atk, this.bdef, this.bspa, this.bspd, this.bspe] =
-      this.pokemonData.stats;
-    this.atk = Math.round(this.pokemonData.stats[1] * atkRatio);
-    this.def = Math.round(this.pokemonData.stats[2] * defRatio);
-    this.spa = Math.round(this.pokemonData.stats[3] * spaRatio);
-    this.spd = Math.round(this.pokemonData.stats[4] * spdRatio);
-    this.spe = Math.round(this.pokemonData.stats[5] * speRatio);
-
-    // set moves and ability
-    this.addMoves(this.pokemonData);
-    this.setAbility(this.pokemonData);
-    this.applyAbility();
-
-    this.battle.addToLog(`${oldName} transformed into ${this.name}!`);
-  }
-
-  useMove(moveId, targetPokemonId) {
-    // make sure pokemon can move
-    if (!this.canMove()) {
-      return;
-    }
-    // make sure move exists and is not on cooldown & disabled
-    const moveData = getMove(moveId);
-    if (
-      !moveData ||
-      this.moveIds[moveId].cooldown > 0 ||
-      this.moveIds[moveId].disabled
-    ) {
-      return;
-    }
-    // check to see if target is valid
-    const primaryTarget = this.battle.allPokemon[targetPokemonId];
-    if (
-      !primaryTarget ||
-      !this.battle.getEligibleTargets(this, moveId).includes(primaryTarget)
-    ) {
-      return;
-    }
-
-    // clear battle log
-    this.battle.clearLog();
-
-    // reset combat readiness
-    this.combatReadiness = 0;
-
-    let canUseMove = true;
-
-    // check status conditions
-    if (this.status.statusId !== null) {
-      switch (this.status.statusId) {
-        case statusConditions.FREEZE:
-          // if move is flare blitz, thaw out
-          if (moveId === "m394" || moveId === "m394-1") {
-            this.removeStatus();
-            break;
-          }
-
-          // thaw chance (turns => chance): 0 => 0%, 1 => 40%, 2 => 80%, 3 => 100%
-          const thawChance = this.status.turns * 0.4;
-          const thawRoll = Math.random();
-          if (thawRoll < thawChance) {
-            this.removeStatus();
-          } else {
-            this.battle.addToLog(`${this.name} is frozen solid!`);
-            canUseMove = false;
-          }
-          break;
-        case statusConditions.PARALYSIS:
-          // 25% chance to be paralyzed
-          const paralysisRoll = Math.random();
-          if (paralysisRoll < 0.25) {
-            this.battle.addToLog(`${this.name} is paralyzed and can't move!`);
-            canUseMove = false;
-          }
-          break;
-        case statusConditions.SLEEP:
-          // ignore if sleep talk is used
-          if (moveId === "m214") {
-            break;
-          }
-          // sleep wakeup chance: 0 turns: 0%, 1 turn: 66%, 2 turns: 100%
-          const wakeupChance = this.status.turns * 0.66;
-          const wakeupRoll = Math.random();
-          if (wakeupRoll < wakeupChance) {
-            this.removeStatus();
-          } else {
-            this.battle.addToLog(`${this.name} is fast asleep!`);
-            canUseMove = false;
-          }
-          break;
-        default:
-          break;
-      }
-    }
-
-    if (canUseMove) {
-      const eventArgs = {
-        canUseMove,
-        source: this,
-        primaryTarget,
-        moveId,
-      };
-
-      // trigger before move events
-      this.battle.eventHandler.emit(battleEventEnum.BEFORE_MOVE, eventArgs);
-
-      canUseMove = eventArgs.canUseMove;
-    }
-
-    // check if pokemon can use move make sure pokemon is alive
-    canUseMove = !!(canUseMove && !this.isFainted);
-
-    // get move data and execute move
-    if (canUseMove) {
-      // see if move log should be silenced
-      const isSilenced =
-        moveData.silenceIf && moveData.silenceIf(this.battle, this);
-      // if pokemon alive, get all targets
-      const allTargets = this.getTargets(moveId, targetPokemonId);
-      if (!isSilenced) {
-        const targetString =
-          moveData.targetPattern === targetPatterns.ALL ||
-          moveData.targetPattern === targetPatterns.ALL_EXCEPT_SELF ||
-          moveData.targetPattern === targetPatterns.RANDOM ||
-          moveData.targetPosition === targetPositions.SELF
-            ? "!"
-            : ` against ${primaryTarget.name}!`;
-        this.battle.addToLog(
-          `${this.name} used ${moveData.name}${targetString}`
-        );
-      }
-      // calculate miss
-      const missedTargets = this.getMisses(moveId, allTargets);
-      // if misses, log miss
-      if (missedTargets.length > 0 && !isSilenced) {
-        this.battle.addToLog(
-          `Missed ${missedTargets.map((target) => target.name).join(", ")}!`
-        );
-      }
-
-      // set cooldown
-      this.moveIds[moveId].cooldown = moveData.cooldown;
-
-      // trigger before execute move events
-      const executeEventArgs = {
-        source: this,
-        primaryTarget,
-        targets: allTargets,
-        missedTargets,
-        moveId,
-      };
-      this.battle.eventHandler.emit(
-        battleEventEnum.BEFORE_MOVE_EXECUTE,
-        executeEventArgs
-      );
-
-      // HACK: set primary / all targets for later damage calculation
-      this.currentPrimaryTarget = primaryTarget;
-      this.currentAllTargets = allTargets;
-
-      // execute move
-      this.executeMove({
-        moveId,
-        primaryTarget,
-        allTargets,
-        missedTargets,
-      });
-
-      // clear primary / all targets
-      this.currentPrimaryTarget = null;
-      this.currentAllTargets = null;
-
-      // after move event
-      const eventArgs = {
-        source: this,
-        primaryTarget,
-        targets: allTargets,
-        missedTargets,
-        moveId,
-      };
-      this.battle.eventHandler.emit(battleEventEnum.AFTER_MOVE, eventArgs);
-    }
-
-    // end turn
-    this.battle.nextTurn();
-  }
-
-  /**
-   * @param {object} param0
-   * @param {MoveIdEnum} param0.moveId
-   * @param {object} param0.primaryTarget
-   * @param {Array<object>} param0.allTargets
-   * @param {Array<object>=} param0.missedTargets
-   */
-  executeMove({ moveId, primaryTarget, allTargets, missedTargets = [] }) {
-    const move = getMove(moveId);
-    if (!move) {
-      logger.error(`Move ${moveId} not found.`);
-      return;
-    }
-
-    if (!move.isLegacyMove) {
-      move.execute({
-        battle: this.battle,
-        source: this,
-        primaryTarget,
-        allTargets,
-        missedTargets,
-      });
-    } else {
-      const legacyMove = /** @type {any} */ (move);
-      legacyMove.execute(
-        this.battle,
-        this,
-        primaryTarget,
-        allTargets,
-        missedTargets
-      );
-    }
-  }
-
-  /**
-   * @param {object} param0
-   * @param {Move} param0.move
-   * @param {Pokemon} param0.target
-   * @param {Pokemon} param0.primaryTarget
-   * @param {Array<Pokemon>} param0.allTargets
-   * @param {Array<Pokemon>=} param0.missedTargets
-   * @param {DamageTypeEnum=} param0.atkDamageTypeOverride
-   * @param {number=} param0.attackOverride
-   * @param {DamageTypeEnum=} param0.defDamageTypeOverride
-   * @param {number=} param0.defenseOverride
-   * @param {number=} param0.powerOverride
-   * @param {number=} param0.typeEffectivenessOverride
-   * @param {PokemonTypeEnum=} param0.moveTypeOverride
-   * @param {number=} param0.offTargetDamageMultiplier
-   * @param {number=} param0.missedTargetDamageMultiplier
-   * @returns {number}
-   */
-  calculateMoveDamage({
-    move,
-    target,
-    primaryTarget,
-    allTargets,
-    missedTargets = [],
-    atkDamageTypeOverride = null,
-    attackOverride = null,
-    defDamageTypeOverride = null,
-    defenseOverride = null,
-    powerOverride = null,
-    typeEffectivenessOverride = null,
-    moveTypeOverride = null,
-    offTargetDamageMultiplier = 0.8,
-    missedTargetDamageMultiplier = 0.7,
-  }) {
-    const power = powerOverride || move.power;
-    const { level } = this;
-    const atkDamageType = atkDamageTypeOverride || move.damageType;
-    const attackRaw =
-      attackOverride ||
-      (atkDamageType === damageTypes.PHYSICAL ? this.atk : this.spa);
-    // modify attack amount -- any attack over 800 is only half as effective
-    const attack = attackRaw > 800 ? 800 + (attackRaw - 800) / 2 : attackRaw;
-
-    const defDamageType = defDamageTypeOverride || move.damageType;
-    const defense =
-      defenseOverride ||
-      (defDamageType === damageTypes.PHYSICAL
-        ? target.getDef()
-        : target.getSpd());
-    const stab = this.type1 === move.type || this.type2 === move.type ? 1.5 : 1;
-    const missMult = missedTargets.includes(target) ? 0.7 : 1;
-    const moveType = moveTypeOverride || move.type;
-    const typeEffectiveness =
-      typeEffectivenessOverride !== null
-        ? typeEffectivenessOverride
-        : this.getTypeDamageMultiplier(moveType, target);
-    // balance type
-    let typeMultiplier = 1;
-    if (typeEffectiveness >= 4) {
-      typeMultiplier = 2;
-    } else if (typeEffectiveness >= 2) {
-      typeMultiplier = 1.5;
-    } else if (typeEffectiveness >= 1) {
-      typeMultiplier = 1;
-    } else if (typeEffectiveness >= 0.5) {
-      typeMultiplier = 0.75;
-    } else if (typeEffectiveness >= 0.25) {
-      typeMultiplier = 0.5;
-    } else {
-      typeMultiplier = 0.35;
-    }
-    const random = Math.random() * (1 - 0.85) + 0.85;
-    const burn = this.status.statusId === statusConditions.BURN ? 0.65 : 1;
-
-    let weatherMult = 1;
-    if (!this.battle.isWeatherNegated()) {
-      const { weather } = this.battle;
-      if (weather.weatherId === weatherConditions.SUN) {
-        if (move.type === pokemonTypes.FIRE) {
-          weatherMult = 1.5;
-        } else if (move.type === pokemonTypes.WATER) {
-          weatherMult = 0.5;
-        }
-      } else if (weather.weatherId === weatherConditions.RAIN) {
-        if (move.type === pokemonTypes.FIRE) {
-          weatherMult = 0.5;
-        } else if (move.type === pokemonTypes.WATER) {
-          weatherMult = 1.5;
-        }
-      }
-    }
-
-    const maybeOffTargetMult =
-      target === primaryTarget ? 1 : offTargetDamageMultiplier;
-
-    /* console.log("power", power)
-      console.log("level", level)
-      console.log("attack", attack)
-      console.log("defense", defense)
-      console.log("stab", stab)
-      console.log("type", type)
-      console.log("random", random) */
-
-    const damage = Math.floor(
-      ((((2 * level) / 5 + 2) * power * attack) / defense / 50 + 2) *
-        stab *
-        typeMultiplier *
-        random *
-        burn *
-        weatherMult *
-        missMult *
-        maybeOffTargetMult
-    );
-
-    return Math.max(damage, 1);
-  }
-
-  /**
-   * Called if the Pokemon is the active Pokemon but can't move.
-   * @returns null
-   */
-  skipTurn() {
-    // make sure its this Pokemon's turn
-    if (this.battle.activePokemon !== this) {
-      return;
-    }
-    // make sure Pokemon can't move
-    if (this.canMove()) {
-      return;
-    }
-
-    // clear battle log
-    this.battle.clearLog();
-
-    // reset combat readiness
-    this.combatReadiness = 0;
-
-    // check pre-move status conditions that tick regardless of move
-    if (this.status.statusId !== null) {
-      switch (this.status.statusId) {
-        case statusConditions.FREEZE:
-          // thaw chance (turns => chance): 0 => 0%, 1 => 40%, 2 => 80%, 3 => 100%
-          const thawChance = this.status.turns * 0.4;
-          const thawRoll = Math.random();
-          if (thawRoll < thawChance) {
-            this.removeStatus();
-          }
-          break;
-        case statusConditions.SLEEP:
-          // sleep wakeup chance: 0 turns: 0%, 1 turn: 66%, 2 turns: 100%
-          const wakeupChance = this.status.turns * 0.66;
-          const wakeupRoll = Math.random();
-          if (wakeupRoll < wakeupChance) {
-            this.removeStatus();
-          }
-          break;
-        default:
-          break;
-      }
-    }
-
-    // end turn
-    this.battle.nextTurn();
-  }
-
-  /**
-   * Check that the Pokemon may take a valid action.
-   * @returns True or False
-   */
-  canMove() {
-    // make sure its this pokemon's turn
-    if (this.battle.activePokemon !== this) {
-      return false;
-    }
-
-    // if fainted, return
-    if (this.isFainted) {
-      return false;
-    }
-
-    // if incapacitated, return
-    if (this.incapacitated) {
-      return false;
-    }
-
-    // for all non-disabled moves not on cooldown, check to see if valid targets exist
-    for (const moveId in this.moveIds) {
-      if (this.moveIds[moveId].disabled || this.moveIds[moveId].cooldown > 0) {
-        continue;
-      }
-      const eligibleTargets = this.battle.getEligibleTargets(this, moveId);
-      if (eligibleTargets.length > 0) {
-        return true;
-      }
-    }
-
-    // if no valid targets exist, return false
-    return false;
-  }
-
-  getTypeDamageMultiplier(moveType, targetPokemon) {
-    let mult = 1;
-    if (typeAdvantages[moveType]) {
-      let adv = typeAdvantages[moveType][targetPokemon.type1];
-      if (adv !== undefined) {
-        mult *= adv;
-      }
-
-      adv = typeAdvantages[moveType][targetPokemon.type2];
-      if (adv !== undefined) {
-        mult *= adv;
-      }
-    }
-
-    const eventArgs = {
-      source: this,
-      target: targetPokemon,
-      moveType,
-      multiplier: mult,
-    };
-    this.battle.eventHandler.emit(
-      battleEventEnum.CALCULATE_TYPE_MULTIPLIER,
-      eventArgs
-    );
-
-    return eventArgs.multiplier;
-  }
-
-  getPatternTargets(targetParty, targetPattern, targetPosition, moveId = null) {
-    const targetRow = Math.floor((targetPosition - 1) / targetParty.cols);
-    const targetCol = (targetPosition - 1) % targetParty.cols;
-    const targets = [];
-
-    switch (targetPattern) {
-      case targetPatterns.ALL:
-        // return all pokemon in party
-        for (const pokemon of targetParty.pokemons) {
-          if (this.battle.isPokemonHittable(pokemon, moveId)) {
-            targets.push(pokemon);
-          }
-        }
-        break;
-      case targetPatterns.ALL_EXCEPT_SELF:
-        // return all pokemon in party except self
-        for (const pokemon of targetParty.pokemons) {
-          if (
-            this.battle.isPokemonHittable(pokemon, moveId) &&
-            pokemon !== this
-          ) {
-            targets.push(pokemon);
-          }
-        }
-        break;
-      case targetPatterns.COLUMN:
-        // return all pokemon in column
-        for (const pokemon of targetParty.pokemons) {
-          if (
-            this.battle.isPokemonHittable(pokemon, moveId) &&
-            (pokemon.position - 1) % targetParty.cols === targetCol
-          ) {
-            targets.push(pokemon);
-          }
-        }
-        break;
-      case targetPatterns.ROW:
-        // return all pokemon in row
-        for (const pokemon of targetParty.pokemons) {
-          if (
-            this.battle.isPokemonHittable(pokemon, moveId) &&
-            Math.floor((pokemon.position - 1) / targetParty.cols) === targetRow
-          ) {
-            targets.push(pokemon);
-          }
-        }
-        break;
-      case targetPatterns.RANDOM:
-        // return random pokemon in party
-        const validPokemons = [];
-        for (const pokemon of targetParty.pokemons) {
-          if (this.battle.isPokemonHittable(pokemon, moveId)) {
-            validPokemons.push(pokemon);
-          }
-        }
-        targets.push(
-          validPokemons[Math.floor(Math.random() * validPokemons.length)]
-        );
-        break;
-      case targetPatterns.SQUARE:
-        // if row index or column index within 1 of target, add to targets
-        for (const pokemon of targetParty.pokemons) {
-          if (this.battle.isPokemonHittable(pokemon, moveId)) {
-            const pokemonRow = Math.floor(
-              (pokemon.position - 1) / targetParty.cols
-            );
-            const pokemonCol = (pokemon.position - 1) % targetParty.cols;
-            if (
-              Math.abs(targetRow - pokemonRow) <= 1 &&
-              Math.abs(targetCol - pokemonCol) <= 1
-            ) {
-              targets.push(pokemon);
-            }
-          }
-        }
-        break;
-      case targetPatterns.CROSS:
-        // target manhattan distance <= 1, add to targets
-        for (const pokemon of targetParty.pokemons) {
-          if (this.battle.isPokemonHittable(pokemon, moveId)) {
-            const pokemonRow = Math.floor(
-              (pokemon.position - 1) / targetParty.cols
-            );
-            const pokemonCol = (pokemon.position - 1) % targetParty.cols;
-            if (
-              Math.abs(targetRow - pokemonRow) +
-                Math.abs(targetCol - pokemonCol) <=
-              1
-            ) {
-              targets.push(pokemon);
-            }
-          }
-        }
-        break;
-      default:
-        // default is single
-
-        // get target pokemon
-        const targetPokemon =
-          targetParty.pokemons[targetRow * targetParty.cols + targetCol];
-        if (this.battle.isPokemonHittable(targetPokemon, moveId)) {
-          targets.push(targetPokemon);
-        }
-
-        break;
-    }
-
-    return targets;
-  }
-
-  getTargets(moveId, targetPokemonId) {
-    const moveData = getMove(moveId);
-    // make sure target exists and is alive
-    const target = this.battle.allPokemon[targetPokemonId];
-    if (!target) {
-      return [];
-    }
-
-    // get party of target
-    const targetParty = this.battle.parties[target.teamName];
-
-    const allTargets = [];
-    return [
-      ...allTargets,
-      ...this.getPatternTargets(
-        targetParty,
-        moveData.targetPattern,
-        target.position,
-        moveId
-      ),
-    ];
-  }
-
-  getMisses(moveId, targetPokemons) {
-    const moveData = getMove(moveId);
-    const misses = [];
-    if (!moveData.accuracy) {
-      return misses;
-    }
-    for (const target of targetPokemons) {
-      let hitChance =
-        (moveData.accuracy * calculateEffectiveAccuracy(this.acc)) /
-        calculateEffectiveEvasion(target.eva);
-      const damageMult = this.getTypeDamageMultiplier(moveData.type, target);
-      if (damageMult >= 4) {
-        hitChance *= 1.4;
-      } else if (damageMult >= 2) {
-        hitChance *= 1.15;
-      } else if (damageMult === 0) {
-        hitChance = 0;
-      } else if (damageMult <= 0.25) {
-        hitChance *= 0.6;
-      } else if (damageMult <= 0.5) {
-        hitChance *= 0.8;
-      }
-
-      // weather check
-      if (!this.battle.isWeatherNegated()) {
-        if (this.battle.weather.weatherId === weatherConditions.SUN) {
-          if (
-            moveId === "m87" ||
-            moveId === "m87-1" ||
-            moveId === "m542" ||
-            moveId === "m542-1"
-          ) {
-            hitChance *= 0.75;
-          }
-        } else if (this.battle.weather.weatherId === weatherConditions.RAIN) {
-          if (
-            moveId === "m87" ||
-            moveId === "m87-1" ||
-            moveId === "m542" ||
-            moveId === "m542-1"
-          ) {
-            hitChance = 150;
-          }
-        } else if (this.battle.weather.weatherId === weatherConditions.HAIL) {
-          if (moveId === "m59") {
-            hitChance = 150;
-          }
-        }
-      }
-
-      const calculateMissArgs = {
-        target,
-        hitChance,
-        source: this,
-      };
-      this.battle.eventHandler.emit(
-        battleEventEnum.CALCULATE_MISS,
-        calculateMissArgs
-      );
-
-      if (Math.random() > calculateMissArgs.hitChance / 100) {
-        misses.push(target);
-      }
-    }
-    return misses;
-  }
-
-  switchUsers(userId) {
-    if (this.userId === userId) {
-      return false;
-    }
-    // get user info from battle
-    const user = this.battle.users[userId];
-    if (!user) {
-      return false;
-    }
-    const { teamName } = user;
-    if (!teamName) {
-      return false;
-    }
-
-    // set teamName and userId
-    this.teamName = teamName;
-    this.userId = userId;
-
-    this.battle.addToLog(`${this.name} switched to ${user.username}'s team!`);
-
-    return true;
-  }
-
-  // eslint-disable-next-line no-unused-vars
-  switchPositions(userId, newPosition, source) {
-    const currentIndex = this.position - 1;
-    const newIndex = newPosition - 1;
-
-    // get user info from battle
-    const user = this.battle.users[userId];
-    if (!user) {
-      return false;
-    }
-    const newTeamName = user.teamName;
-    if (!newTeamName) {
-      return false;
-    }
-
-    const oldParty = this.battle.parties[this.teamName];
-    const newParty = this.battle.parties[newTeamName];
-    if (!newParty) {
-      return false;
-    }
-
-    // check if position is valid
-    if (newPosition < 1 || newPosition > newParty.pokemons.length) {
-      return false;
-    }
-    // if pokemon exists in new position, return false
-    if (newParty.pokemons[newIndex]) {
-      return false;
-    }
-
-    // if new team, attempt to switch teams
-    if (this.teamName !== newTeamName) {
-      if (!this.switchUsers(userId)) {
-        return false;
-      }
-    }
-
-    // remove from old position
-    oldParty.pokemons[currentIndex] = null;
-
-    // add to new position
-    newParty.pokemons[newIndex] = this;
-    this.position = newPosition;
-
-    this.battle.addToLog(
-      `${this.name} switched to position to ${newPosition}!`
-    );
-
-    return true;
-  }
-
-  dealDamage(damage, target, damageInfo) {
-    if (damage <= 0) {
-      return 0;
-    }
-
-    // if pvp, deal 15% less damage
-    if (this.battle.isPvp) {
-      damage = Math.max(1, Math.floor(damage * 0.85));
-    }
-
-    const eventArgs = {
-      target,
-      damage,
-      source: this,
-      damageInfo,
-    };
-
-    this.battle.eventHandler.emit(
-      battleEventEnum.BEFORE_DAMAGE_DEALT,
-      eventArgs
-    );
-    damage = eventArgs.damage;
-
-    const damageDealt = target.takeDamage(damage, this, damageInfo);
-
-    if (damageDealt > 0) {
-      const afterDamageArgs = {
-        target,
-        damage: damageDealt,
-        source: this,
-        damageInfo,
-      };
-      this.battle.eventHandler.emit(
-        battleEventEnum.AFTER_DAMAGE_DEALT,
-        afterDamageArgs
-      );
-    }
-
-    return damageDealt;
-  }
-
-  takeDamage(damage, source, damageInfo) {
-    if (this.isFainted) {
-      return 0;
-    }
-
-    // if frozen and fire type or scald, thaw and deal 1.5x damage
-    const freezeCheck =
-      this.status.statusId === statusConditions.FREEZE &&
-      damageInfo.type === "move" &&
-      getMove(damageInfo.moveId) !== undefined &&
-      (getMove(damageInfo.moveId).type === pokemonTypes.FIRE ||
-        damageInfo.moveId === "m503");
-    if (freezeCheck) {
-      if (this.removeStatus()) {
-        damage = Math.floor(damage * 1.5);
-      }
-    }
-
-    // if shield, take shield damage and return 0
-    const shieldData = this.effectIds.shield;
-    if (shieldData && shieldData.args && shieldData.args.shield) {
-      const shieldDamage = Math.min(damage, shieldData.args.shield);
-      shieldData.args.shield -= shieldDamage;
-
-      this.battle.addToLog(
-        `${this.name} took ${shieldDamage} shield damage! (${shieldData.args.shield} left)`
-      );
-      if (shieldData.args.shield <= 0) {
-        this.removeEffect("shield");
-      }
-      return 0;
-    }
-
-    const eventArgs = {
-      target: this,
-      damage,
-      source,
-      damageInfo,
-      maxDamage: Number.MAX_SAFE_INTEGER,
-    };
-
-    this.battle.eventHandler.emit(
-      battleEventEnum.BEFORE_DAMAGE_TAKEN,
-      eventArgs
-    );
-    damage = Math.min(eventArgs.damage, eventArgs.maxDamage);
-
-    const oldHp = this.hp;
-    if (oldHp <= 0 || this.isFainted) {
-      return 0;
-    }
-    this.hp = Math.max(0, this.hp - damage);
-    const damageTaken = oldHp - this.hp;
-    this.battle.addToLog(`${this.name} took ${damageTaken} damage!`);
-    if (this.hp <= 0) {
-      this.hp = 0;
-      this.faint(source);
-    }
-
-    if (damageTaken > 0) {
-      const afterDamageArgs = {
-        target: this,
-        damage: damageTaken,
-        source,
-        damageInfo,
-      };
-      this.battle.eventHandler.emit(
-        battleEventEnum.AFTER_DAMAGE_TAKEN,
-        afterDamageArgs
-      );
-    }
-
-    return damageTaken;
-  }
-
-  takeFaint(source) {
-    if (this.isFainted) {
-      return;
-    }
-    // trigger before cause faint effects
-    const beforeCauseFaintArgs = {
-      target: this,
-      source,
-      canFaint: true,
-    };
-    this.battle.eventHandler.emit(
-      battleEventEnum.BEFORE_CAUSE_FAINT,
-      beforeCauseFaintArgs
-    );
-    if (!beforeCauseFaintArgs.canFaint) {
-      return;
-    }
-
-    this.faint(source);
-  }
-
-  faint(source) {
-    // TODO: trigger before faint effects
-
-    this.hp = 0;
-    this.isFainted = true;
-    this.removeAbility();
-    this.battle.addToLog(`${this.name} fainted!`);
-
-    // trigger after faint effects
-    const afterFaintArgs = {
-      target: this,
-      source,
-    };
-    this.battle.eventHandler.emit(battleEventEnum.AFTER_FAINT, afterFaintArgs);
-  }
-
-  // eslint-disable-next-line no-unused-vars
-  beRevived(reviveHp, source) {
-    if (!this.isFainted) {
-      return;
-    }
-    if (reviveHp <= 0) {
-      return;
-    }
-
-    this.hp = Math.min(this.maxHp, reviveHp);
-    this.isFainted = false;
-    this.battle.addToLog(`${this.name} was revived!`);
-
-    // re-add ability
-    this.applyAbility();
-  }
-
-  applyAbility() {
-    const { abilityId } = this.ability;
-    const abilityData = getAbility(/** @type {AbilityIdEnum} */ (abilityId));
-    if (!abilityData || !abilityData.abilityAdd) {
-      // TODO: not all abilities are implemented
-      // logger.error(`Ability ${abilityId} does not exist.`);
-      return;
-    }
-
-    if (!abilityData.isLegacyAbility) {
-      this.ability.data = abilityData.abilityAdd({
-        battle: this.battle,
-        source: this,
-        target: this,
-      });
-    } else {
-      const legacyAbility = /** @type {any} */ (abilityData);
-      legacyAbility.abilityAdd(this.battle, this, this);
-    }
-    this.ability.applied = true;
-  }
-
-  removeAbility() {
-    // remove ability effects
-    const { abilityId, applied } = this.ability;
-    const abilityData = getAbility(/** @type {AbilityIdEnum} */ (abilityId));
-    if (!abilityData || !abilityData.abilityRemove) {
-      // TODO: not all abilities are implemented
-      // logger.error(`Ability ${abilityId} does not exist.`);
-      return;
-    }
-    if (!abilityId || !applied) {
-      logger.error(
-        `Ability ${abilityId} is not applied to Pokemon ${this.id} ${this.name}.`
-      );
-      return;
-    }
-
-    if (!abilityData.isLegacyAbility) {
-      abilityData.abilityRemove({
-        battle: this.battle,
-        source: this,
-        target: this,
-        properties: this.ability.data,
-      });
-    } else {
-      const legacyAbility = /** @type {any} */ (abilityData);
-      legacyAbility.abilityRemove(this.battle, this, this);
-    }
-  }
-
-  giveHeal(heal, target, healInfo) {
-    const healGiven = target.takeHeal(heal, this, healInfo);
-    return healGiven;
-  }
-
-  // eslint-disable-next-line no-unused-vars
-  takeHeal(heal, source, healInfo) {
-    const oldHp = this.hp;
-    if (oldHp <= 0 || this.isFainted) {
-      return 0;
-    }
-    this.hp = Math.min(this.maxHp, this.hp + heal);
-    const healTaken = this.hp - oldHp;
-    if (healTaken > 0) {
-      this.battle.addToLog(`${this.name} healed ${healTaken} HP!`);
-    }
-    return healTaken;
-  }
-
-  boostCombatReadiness(source, amount, triggerEvents = true) {
-    // if faint, do nothing
-    if (this.isFainted) {
-      return 0;
-    }
-
-    if (this.restricted) {
-      this.battle.addToLog(
-        `${this.name} is restricted and cannot gain combat readiness!`
-      );
-      return 0;
-    }
-
-    const beforeBoostArgs = {
-      target: this,
-      source,
-      amount,
-    };
-    if (triggerEvents) {
-      this.battle.eventHandler.emit(
-        battleEventEnum.BEFORE_CR_GAINED,
-        beforeBoostArgs
-      );
-    }
-
-    const oldCombatReadiness = this.combatReadiness;
-    this.combatReadiness = Math.min(
-      100,
-      this.combatReadiness + beforeBoostArgs.amount
-    );
-    const combatReadinessGained = this.combatReadiness - oldCombatReadiness;
-    this.battle.addToLog(
-      `${this.name} gained ${Math.round(
-        combatReadinessGained
-      )} combat readiness!`
-    );
-
-    if (combatReadinessGained > 0 && triggerEvents) {
-      const eventArgs = {
-        target: this,
-        source,
-        combatReadinessGained: amount,
-      };
-      this.battle.eventHandler.emit(battleEventEnum.AFTER_CR_GAINED, eventArgs);
-    }
-
-    return combatReadinessGained;
-  }
-
-  reduceCombatReadiness(source, amount) {
-    // if faint, do nothing
-    if (this.isFainted) {
-      return 0;
-    }
-
-    const oldCombatReadiness = this.combatReadiness;
-    this.combatReadiness = Math.max(0, this.combatReadiness - amount);
-    const combatReadinessLost = oldCombatReadiness - this.combatReadiness;
-    this.battle.addToLog(
-      `${this.name} lost ${Math.round(combatReadinessLost)} combat readiness!`
-    );
-    return combatReadinessLost;
-  }
-
-  /**
-   * @template {EffectIdEnum} K
-   * @param {K} effectId
-   * @param {number} duration
-   * @param {BattlePokemon} source
-   * @param {EffectInitialArgsTypeFromId<K>} initialArgs
-   */
-  applyEffect(effectId, duration, source, initialArgs) {
-    // if faint, do nothing
-    if (this.isFainted) {
-      return;
-    }
-    const effect = getEffect(effectId);
-    if (!effect) {
-      logger.error(`Effect ${effectId} does not exist.`);
-      return;
-    }
-
-    duration = this.battle.activePokemon === this ? duration + 1 : duration;
-
-    // if effect already exists for longer or equal duration, do nothing (special case for shield)
-    if (
-      this.effectIds[effectId] &&
-      this.effectIds[effectId].duration >= duration &&
-      effectId !== "shield"
-    ) {
-      return;
-    }
-
-    // if effect exists, refresh duration
-    // TODO: should this be modified?
-    if (this.effectIds[effectId]) {
-      duration = Math.max(this.effectIds[effectId].duration, duration);
-      this.effectIds[effectId].duration = duration;
-      // shield special case
-      if (effectId !== "shield") {
-        return;
-      }
-    }
-
-    // trigger before effect add events
-    const beforeAddArgs = {
-      target: this,
-      source,
-      effectId,
-      duration,
-      initialArgs,
-      canAdd: true,
-    };
-    this.battle.eventHandler.emit(
-      battleEventEnum.BEFORE_EFFECT_ADD,
-      beforeAddArgs
-    );
-    if (!beforeAddArgs.canAdd) {
-      return;
-    }
-    duration = beforeAddArgs.duration;
-
-    // special case for shield
-    let oldShield = {};
-    if (effectId === "shield") {
-      // if shield already exists, keep note of it
-      if (this.effectIds[effectId]) {
-        oldShield = this.effectIds[effectId].args;
-      }
-    }
-    this.effectIds[effectId] = {
-      duration,
-      source,
-      initialArgs,
-    };
-    if (oldShield) {
-      this.effectIds[effectId].args = oldShield;
-    }
-
-    if (!effect.isLegacyEffect) {
-      this.effectIds[effectId].args =
-        effect.effectAdd({
-          battle: this.battle,
-          source,
-          target: this,
-          initialArgs,
-        }) || {};
-    } else {
-      const legacyEffect = /** @type {any} */ (effect);
-      this.effectIds[effectId].args =
-        legacyEffect.effectAdd(this.battle, source, this, initialArgs) || {};
-    }
-
-    if (this.effectIds[effectId] !== undefined) {
-      // trigger after add effect events
-      const afterAddArgs = {
-        target: this,
-        source,
-        effectId,
-        duration,
-        initialArgs,
-        args: this.effectIds[effectId].args,
-      };
-      this.battle.eventHandler.emit(
-        battleEventEnum.AFTER_EFFECT_ADD,
-        afterAddArgs
-      );
-    }
-  }
-
-  dispellEffect(effectId) {
-    const effectData = getEffect(effectId);
-
-    // if effect doesn't exist, do nothing
-    if (!this.effectIds[effectId]) {
-      return false;
-    }
-
-    // if effect is not dispellable, do nothing
-    if (!effectData.dispellable) {
-      return false;
-    }
-
-    return this.removeEffect(effectId);
-  }
-
-  /**
-   * @template {EffectIdEnum} K
-   * @param {K} effectId
-   * @returns {{
-   *  duration: number,
-   *  source: Pokemon,
-   *  initialArgs: K extends keyof RegisteredEffects ? EffectInitialArgsType<RegisteredEffects[K]> : any,
-   *  args: K extends keyof RegisteredEffects ? EffectPropertiesType<RegisteredEffects[K]> : any
-   * } | undefined}
-   */
-  getEffectInstance(effectId) {
-    // @ts-ignore
-    return this.effectIds[effectId];
-  }
-
-  /**
-   * Forcefully deletes the effect instance from this Pokemon. ONLY USE to bypass effect removal logic.
-   * @param {EffectIdEnum} effectId
-   */
-  deleteEffectInstance(effectId) {
-    delete this.effectIds[effectId];
-  }
-
-  removeEffect(effectId) {
-    // if effect doesn't exist, do nothing
-    if (!this.effectIds[effectId]) {
-      return false;
-    }
-    const effect = getEffect(effectId);
-    if (!effect) {
-      logger.error(`Effect ${effectId} does not exist.`);
-      return;
-    }
-
-    if (!effect.isLegacyEffect) {
-      // @ts-ignore
-      effect.effectRemove({
-        battle: this.battle,
-        target: this,
-        properties: this.effectIds[effectId].args,
-        initialArgs: this.effectIds[effectId].initialArgs,
-      });
-    } else {
-      const legacyEffect = /** @type {any} */ (effect);
-      legacyEffect.effectRemove(
-        this.battle,
-        this,
-        this.effectIds[effectId].args,
-        this.effectIds[effectId].initialArgs
-      );
-    }
-
-    if (this.effectIds[effectId] !== undefined) {
-      const afterRemoveArgs = {
-        target: this,
-        source: this.effectIds[effectId].source,
-        effectId,
-        duration: this.effectIds[effectId].duration,
-        initialArgs: this.effectIds[effectId].initialArgs,
-        args: this.effectIds[effectId].args,
-      };
-      this.battle.eventHandler.emit(
-        battleEventEnum.AFTER_EFFECT_REMOVE,
-        afterRemoveArgs
-      );
-    }
-
-    delete this.effectIds[effectId];
-    return true;
-  }
-
-  applyStatus(statusId, source, { startingTurns = 0 } = {}) {
-    // if faint, do nothing
-    if (this.isFainted) {
-      return;
-    }
-
-    // if status already exists, do nothing
-    if (this.status.statusId) {
-      this.battle.addToLog(`${this.name} already has a status condition!`);
-      return;
-    }
-
-    // trigger before apply status events
-    const beforeApplyArgs = {
-      target: this,
-      source,
-      statusId,
-      canApply: true,
-    };
-    this.battle.eventHandler.emit(
-      battleEventEnum.BEFORE_STATUS_APPLY,
-      beforeApplyArgs
-    );
-    if (!beforeApplyArgs.canApply) {
-      return;
-    }
-
-    let statusApplied = false;
-    switch (statusId) {
-      // TODO: other status effects
-      case statusConditions.BURN:
-        if (
-          this.type1 === pokemonTypes.FIRE ||
-          this.type2 === pokemonTypes.FIRE
-        ) {
-          this.battle.addToLog(
-            `${this.name}'s Fire type renders it immune to burns!`
-          );
-          break;
-        }
-
-        this.status = {
-          statusId,
-          source,
-          turns: startingTurns,
-        };
-        this.battle.addToLog(`${this.name} was burned!`);
-        statusApplied = true;
-        break;
-      case statusConditions.FREEZE:
-        if (
-          this.type1 === pokemonTypes.ICE ||
-          this.type2 === pokemonTypes.ICE
-        ) {
-          this.battle.addToLog(
-            `${this.name}'s Ice type renders it immune to freezing!`
-          );
-          break;
-        }
-
-        if (this.battle.weather.weatherId === weatherConditions.SUN) {
-          this.battle.addToLog(
-            `${this.name} was protected from freezing by the sun!`
-          );
-          break;
-        }
-
-        this.status = {
-          statusId,
-          source,
-          turns: startingTurns,
-        };
-        this.battle.addToLog(`${this.name} was frozen!`);
-        statusApplied = true;
-        break;
-      case statusConditions.PARALYSIS:
-        if (
-          this.type1 === pokemonTypes.ELECTRIC ||
-          this.type2 === pokemonTypes.ELECTRIC
-        ) {
-          this.battle.addToLog(
-            `${this.name}'s Electric type renders it immune to paralysis!`
-          );
-          break;
-        }
-
-        // reduce speed by 45%
-        this.spe -= Math.floor(this.bspd * 0.45);
-
-        this.status = {
-          statusId,
-          source,
-          turns: startingTurns,
-        };
-        this.battle.addToLog(`${this.name} was paralyzed!`);
-        statusApplied = true;
-        break;
-      case statusConditions.POISON:
-        if (
-          this.type1 === pokemonTypes.POISON ||
-          this.type2 === pokemonTypes.POISON
-        ) {
-          this.battle.addToLog(
-            `${this.name}'s Poison type renders it immune to poison!`
-          );
-          break;
-        }
-
-        this.status = {
-          statusId,
-          source,
-          turns: startingTurns,
-        };
-        this.battle.addToLog(`${this.name} was poisoned!`);
-        statusApplied = true;
-        break;
-      case statusConditions.SLEEP:
-        this.status = {
-          statusId,
-          source,
-          turns: startingTurns,
-        };
-        this.battle.addToLog(`${this.name} fell asleep!`);
-        statusApplied = true;
-        break;
-      case statusConditions.BADLY_POISON:
-        if (
-          this.type1 === pokemonTypes.POISON ||
-          this.type2 === pokemonTypes.POISON
-        ) {
-          this.battle.addToLog(
-            `${this.name}'s Poison type renders it immune to poison!`
-          );
-          break;
-        }
-
-        this.status = {
-          statusId,
-          source,
-          turns: startingTurns,
-        };
-        this.battle.addToLog(`${this.name} was badly poisoned!`);
-        statusApplied = true;
-        break;
-      default:
-        break;
-    }
-
-    if (statusApplied) {
-      const afterStatusArgs = {
-        target: this,
-        source,
-        statusId,
-      };
-      this.battle.eventHandler.emit(
-        battleEventEnum.AFTER_STATUS_APPLY,
-        afterStatusArgs
-      );
-    }
-  }
-
-  tickStatus() {
-    // if status doesn't exist, do nothing
-    if (!this.status.statusId) {
-      return;
-    }
-
-    switch (this.status.statusId) {
-      case statusConditions.POISON:
-        const damage = Math.round(this.maxHp / 6);
-        this.battle.addToLog(`${this.name} is hurt by poison!`);
-        this.takeDamage(damage, this.status.source, {
-          type: "statusCondition",
-          statusId: statusConditions.POISON,
-        });
-        break;
-      case statusConditions.BURN:
-        const burnDamage = Math.round(this.maxHp / 8);
-        this.battle.addToLog(`${this.name} is hurt by its burn!`);
-        this.takeDamage(burnDamage, this.status.source, {
-          type: "statusCondition",
-          statusId: statusConditions.BURN,
-        });
-        break;
-      case statusConditions.BADLY_POISON:
-        const badlyPoisonDamage =
-          Math.round(this.maxHp / 6) * 2 ** this.status.turns;
-        this.battle.addToLog(`${this.name} is hurt by poison!`);
-        this.takeDamage(badlyPoisonDamage, this.status.source, {
-          type: "statusCondition",
-          statusId: statusConditions.BADLY_POISON,
-        });
-        break;
-      default:
-        break;
-    }
-
-    this.status.turns += 1;
-  }
-
-  removeStatus() {
-    // if status doesn't exist, do nothing
-    if (!this.status.statusId) {
-      return false;
-    }
-
-    // TODO: trigger before remove status events
-
-    switch (this.status.statusId) {
-      case statusConditions.BURN:
-        this.battle.addToLog(`${this.name} was cured of its burn!`);
-        break;
-      case statusConditions.FREEZE:
-        this.battle.addToLog(`${this.name} was thawed out!`);
-        break;
-      case statusConditions.PARALYSIS:
-        // restore speed
-        this.spe += Math.floor(this.bspd * 0.45);
-
-        this.battle.addToLog(`${this.name} was cured of its paralysis!`);
-        break;
-      case statusConditions.POISON:
-        this.battle.addToLog(`${this.name} was cured of its poison!`);
-        break;
-      case statusConditions.SLEEP:
-        this.battle.addToLog(`${this.name} woke up!`);
-        break;
-      default:
-        break;
-    }
-
-    this.status = {
-      statusId: null,
-      turns: 0,
-    };
-
-    return true;
-  }
-
-  // eslint-disable-next-line no-unused-vars
-  disableMove(moveId, source) {
-    // check that move exists
-    if (!this.moveIds[moveId]) {
-      return;
-    }
-
-    // if move already disabled, do nothing
-    if (this.moveIds[moveId].disabled) {
-      return;
-    }
-
-    // disable move
-    this.moveIds[moveId].disabled = true;
-    // this.battle.addToLog(`${this.name}'s ${getMove(moveId).name} was disabled!`);
-  }
-
-  // eslint-disable-next-line no-unused-vars
-  enableMove(moveId, source) {
-    // check that move exists
-    if (!this.moveIds[moveId]) {
-      return;
-    }
-
-    // if move not disabled, do nothing
-    if (!this.moveIds[moveId].disabled) {
-      return;
-    }
-
-    // enable move
-    this.moveIds[moveId].disabled = false;
-    // this.battle.addToLog(`${this.name}'s ${getMove(moveId).name} is no longer disabled!`);
-  }
-
-  tickEffectDurations() {
-    for (const effectId in this.effectIds) {
-      this.effectIds[effectId].duration -= 1;
-      if (this.effectIds[effectId].duration <= 0) {
-        this.removeEffect(effectId);
-      }
-    }
-  }
-
-  tickMoveCooldowns() {
-    for (const moveId in this.moveIds) {
-      if (this.moveIds[moveId].cooldown > 0) {
-        this.moveIds[moveId].cooldown -= 1;
-      }
-    }
-  }
-
-  reduceMoveCooldown(moveId, amount, source, silenced = false) {
-    // check that move exists
-    if (!this.moveIds[moveId]) {
-      return;
-    }
-
-    // check that the move is on cooldown
-    if (this.moveIds[moveId].cooldown === 0) {
-      return;
-    }
-
-    // reduce cooldown
-    const oldCooldown = this.moveIds[moveId].cooldown;
-    this.moveIds[moveId].cooldown -= amount;
-    if (this.moveIds[moveId].cooldown < 0) {
-      this.moveIds[moveId].cooldown = 0;
-    }
-    const newCooldown = this.moveIds[moveId].cooldown;
-
-    if (!silenced) {
-      this.battle.addToLog(
-        `${this.name}'s ${getMove(moveId).name}'s cooldown was reduced by ${
-          oldCooldown - newCooldown
-        } turns!`
-      );
-    }
-    return oldCooldown - newCooldown;
-  }
-
-  effectiveSpeed() {
-    return calculateEffectiveSpeed(this.getSpe());
-  }
-
-  getRowAndColumn() {
-    const party = this.battle.parties[this.teamName];
-    const row = Math.floor((this.position - 1) / party.cols);
-    const col = (this.position - 1) % party.cols;
-    return { row, col };
-  }
-
-  getPartyRowColumn() {
-    const party = this.battle.parties[this.teamName];
-    return { party, ...this.getRowAndColumn() };
-  }
-
-  getEnemyParty() {
-    const teamNames = Object.keys(this.battle.parties);
-    const enemyTeamName =
-      teamNames[0] === this.teamName ? teamNames[1] : teamNames[0];
-    return this.battle.parties[enemyTeamName];
-  }
-
-  getDef() {
-    let { def } = this;
-
-    // if hail and ice, def * 1.5
-    if (
-      !this.battle.isWeatherNegated() &&
-      this.battle.weather.weatherId === weatherConditions.HAIL &&
-      (this.type1 === pokemonTypes.ICE || this.type2 === pokemonTypes.ICE)
-    ) {
-      def = Math.floor(def * 1.5);
-    }
-
-    return def;
-  }
-
-  getSpd() {
-    let { spd } = this;
-
-    // if sandstorm and rock, spd * 1.5
-    if (
-      !this.battle.isWeatherNegated() &&
-      this.battle.weather.weatherId === weatherConditions.SANDSTORM &&
-      (this.type1 === pokemonTypes.ROCK || this.type2 === pokemonTypes.ROCK)
-    ) {
-      spd = Math.floor(spd * 1.5);
-    }
-
-    return spd;
-  }
-
-  getSpe() {
-    let { spe } = this;
-
-    if (!this.battle.isWeatherNegated()) {
-      if (this.battle.weather.weatherId === weatherConditions.SUN) {
-        if (this.ability && this.ability.abilityId === "34") {
-          spe = Math.floor(spe * 1.5);
-        }
-      } else if (this.battle.weather.weatherId === weatherConditions.RAIN) {
-        if (this.ability && this.ability.abilityId === "33") {
-          spe = Math.floor(spe * 1.5);
-        }
-      }
-    }
-
-    return spe;
-  }
-}
-
-class Battle {
-  /* TODO: fix
-  moneyMultiplier;
-  expMultiplier;
-  pokemonExpMultiplier;
-  moneyReward;
-  expReward;
-  pokemonExpReward;
-  winCallback;
-  hasStarted;
-  userIds;
-  users;
-  teams;
-  activePokemon;
-  parties;
-  // { weatherId, duration, source }
-  weather;
-  log;
-  eventHandler;
-  turn;
-  winner;
-  ended;
-  minLevel;
-  level;
-  rewards;
-  rewardString;
-  dailyRewards;
-  npcId;
-  difficulty;
-  isPvp; */
-
-  /**
-   * @param {object} param0
-   * @param {number?=} param0.moneyMultiplier
-   * @param {number?=} param0.expMultiplier
-   * @param {number?=} param0.pokemonExpMultiplier
-   * @param {number?=} param0.level
-   * @param {number?=} param0.equipmentLevel
-   * @param {object?=} param0.rewards
-   * @param {string?=} param0.rewardString
-   * @param {object?=} param0.dailyRewards
-   * @param {Function?=} param0.winCallback
-   * @param {Function?=} param0.loseCallback
-   * @param {string?=} param0.npcId
-   * @param {string?=} param0.difficulty
-   * @param {boolean?=} param0.isPvp
-   */
-  constructor({
-    moneyMultiplier = 1,
-    expMultiplier = 1,
-    pokemonExpMultiplier = 0.2,
-    level = null,
-    equipmentLevel = null,
-    rewards = null,
-    rewardString = null,
-    dailyRewards = null,
-    winCallback = null,
-    loseCallback = null,
-    npcId = null,
-    difficulty = null,
-    isPvp = false,
-  } = {}) {
-    // initial
-    this.baseMoney = 100;
-    this.baseExp = 50;
-
-    this.moneyMultiplier = moneyMultiplier;
-    this.expMultiplier = expMultiplier;
-    this.pokemonExpMultiplier = pokemonExpMultiplier;
-    this.moneyReward = 0;
-    this.expReward = 0;
-    this.pokemonExpReward = 0;
-    this.hasStarted = false;
-    this.userIds = [];
-    // map userId to user
-    this.users = {};
-    // map teamName to team
-    this.teams = {};
-    this.activePokemon = null;
-    // map teamName to party
-    this.parties = {};
-    // map pokemonId to pokemon
-    this.allPokemon = {};
-    this.weather = {
-      weatherId: null,
-      duration: 0,
-      source: null,
-    };
-    this.log = [];
-    this.eventHandler = new BattleEventHandler();
-    this.turn = 0;
-    this.winner = null;
-    this.ended = false;
-    if (equipmentLevel) {
-      if (equipmentLevel < 1 || equipmentLevel > 10) {
-        throw new Error("Invalid equipment level");
-      }
-      this.equipmentLevel = equipmentLevel;
-    }
-    if (level) {
-      if (level < 1 || level > 100) {
-        throw new Error("Invalid level");
-      }
-      this.level = level;
-      this.minLevel = level;
-    }
-    this.rewards = rewards;
-    this.rewardString = rewardString;
-    this.dailyRewards = dailyRewards;
-    this.npcId = npcId;
-    this.difficulty = difficulty;
-    this.winCallback = winCallback;
-    this.loseCallback = loseCallback;
-    this.isPvp = isPvp;
-  }
-
-  addTeam(teamName, isNpc) {
-    this.teams[teamName] = {
-      name: teamName,
-      isNpc,
-      userIds: [],
-    };
-  }
-
-  addTrainer(trainer, pokemons, teamName, rows = 3, cols = 4) {
-    // if user already exists, return
-    if (this.users[trainer.userId]) {
-      return;
-    }
-
-    this.teams[teamName].userIds.push(trainer.userId);
-    this.users[trainer.userId] = {
-      teamName,
-      ...trainer.user,
-    };
-    this.userIds.push(trainer.userId);
-
-    this.addPokemons(trainer, pokemons, teamName, rows, cols);
-  }
-
-  addPokemons(trainer, pokemons, teamName, rows, cols) {
-    const partyPokemons = [];
-    for (const pokemonData of pokemons) {
-      if (pokemonData) {
-        const pokemonInstance = new Pokemon(
-          this,
-          trainer,
-          pokemonData,
-          teamName,
-          partyPokemons.length + 1
-        );
-        partyPokemons.push(pokemonInstance);
-        this.allPokemon[pokemonInstance.id] = pokemonInstance;
-      } else {
-        partyPokemons.push(null);
-      }
-    }
-    // TODO: modify if parties can have different num players
-    this.parties[teamName] = {
-      pokemons: partyPokemons,
-      rows,
-      cols,
-    };
-  }
-
-  increaseCombatReadiness() {
-    // get min ticks for a pokemon to be ready
-    const MAX_CR = 100;
-    let minTicks = Number.MAX_SAFE_INTEGER;
-    let minTicksPokemon = null;
-    for (const partyName in this.parties) {
-      const party = this.parties[partyName];
-      for (const pokemon of party.pokemons) {
-        if (pokemon && !pokemon.isFainted) {
-          const requiredCr = MAX_CR - pokemon.combatReadiness;
-          const ticks = requiredCr / pokemon.effectiveSpeed();
-          if (ticks < minTicks) {
-            minTicks = ticks;
-            minTicksPokemon = pokemon;
-          }
-        }
-      }
-    }
-
-    // set active pokemon
-    if (minTicksPokemon) {
-      this.activePokemon = minTicksPokemon;
-    } else {
-      this.activePokemon = null;
-      return;
-    }
-
-    // if no ticks, return
-    if (minTicks === 0) {
-      return;
-    }
-
-    // increase combat readiness for all pokemon
-    for (const partyName in this.parties) {
-      const party = this.parties[partyName];
-      for (const pokemon of party.pokemons) {
-        if (pokemon && !pokemon.isFainted) {
-          pokemon.combatReadiness = Math.min(
-            MAX_CR,
-            pokemon.combatReadiness + pokemon.effectiveSpeed() * minTicks
-          );
-        }
-      }
-    }
-  }
-
-  start() {
-    this.log.push("The battle begins!");
-    this.turn = 0;
-    this.hasStarted = true;
-
-    // sort all pokemon by speed descending
-    const sortedPokemon = Object.values(this.allPokemon);
-    sortedPokemon.sort((a, b) => b.getSpe() - a.getSpe());
-    this.allPokemon = {};
-    for (const pokemon of sortedPokemon) {
-      this.allPokemon[pokemon.id] = pokemon;
-    }
-
-    // add all abilities
-    Object.entries(this.allPokemon).forEach(([, pokemon]) => {
-      if (pokemon.ability.applied) {
-        return;
-      }
-      pokemon.applyAbility();
-    });
-
-    this.eventHandler.emit(battleEventEnum.BATTLE_BEGIN, {
-      battle: this,
-    });
-
-    // begin turn
-    this.beginTurn();
-  }
-
-  beginTurn() {
-    // push cr
-    this.increaseCombatReadiness();
-
-    // tick move cooldowns
-    if (!this.activePokemon.isFainted) {
-      this.activePokemon.tickMoveCooldowns();
-    }
-
-    // begin turn
-    this.eventHandler.emit(battleEventEnum.TURN_BEGIN);
-
-    // log
-    const userIsNpc = this.isNpc(this.activePokemon.userId);
-    const userString = userIsNpc
-      ? this.users[this.activePokemon.userId].username
-      : `<@${this.activePokemon.userId}>`;
-    if (this.activePokemon.canMove()) {
-      this.log.push(
-        `**[Turn ${this.turn}] It is ${userString}'s ${this.activePokemon.name}'s turn.**`
-      );
-    } else {
-      this.log.push(
-        `**[Turn ${this.turn}] ${userString}'s ${this.activePokemon.name} is unable to move.**`
-      );
-    }
-  }
-
-  nextTurn() {
-    // end turn logic
-    this.emitEvent(battleEventEnum.TURN_END, {
-      activePokemon: this.activePokemon,
-    });
-
-    // tick status effects
-    if (!this.activePokemon.isFainted) {
-      this.activePokemon.tickStatus();
-    }
-
-    // tick effects
-    if (!this.activePokemon.isFainted) {
-      this.activePokemon.tickEffectDurations();
-    }
-
-    // tick weather
-    if (this.weather) {
-      this.tickWeather();
-    }
-
-    // increase turn and check for game end
-    this.turn += 1;
-    if (this.turn > 100) {
-      return this.endBattle();
-    }
-    const undefeatedTeams = [];
-    for (const teamName in this.teams) {
-      let isTeamDefeated = true;
-      for (const pokemon of this.parties[teamName].pokemons) {
-        if (pokemon && !pokemon.isFainted) {
-          isTeamDefeated = false;
-          break;
-        }
-      }
-      if (!isTeamDefeated) {
-        undefeatedTeams.push(teamName);
-      }
-    }
-    if (undefeatedTeams.length === 1) {
-      [this.winner] = undefeatedTeams;
-      return this.endBattle();
-    }
-    if (undefeatedTeams.length === 0) {
-      return this.endBattle();
-    }
-
-    // begin turn
-    this.beginTurn();
-  }
-
-  endBattle() {
-    if (this.winner) {
-      // get loser, and if any loser has next phase, initialize next phase.
-      const loser =
-        Object.keys(this.teams)[0] === this.winner
-          ? Object.keys(this.teams)[1]
-          : Object.keys(this.teams)[0];
-      let nextPhase = false;
-      for (const userId of this.teams[loser].userIds) {
-        const user = this.users[userId];
-        if (user.nextPhase !== undefined) {
-          const userNextPhase = user.nextPhase(this);
-          if (!userNextPhase) {
-            continue;
-          }
-          const { trainer, pokemons, rows, cols } = userNextPhase;
-
-          // clear out all of user's pokemon
-          for (const [pokemonId, pokemon] of Object.entries(this.allPokemon)) {
-            if (pokemon.userId === userId) {
-              delete this.allPokemon[pokemonId];
-            }
-          }
-          // add new pokemon
-          this.addPokemons(trainer, pokemons, loser, rows, cols);
-          nextPhase = true;
-        }
-      }
-      if (nextPhase) {
-        this.addToLog(`**[Phase Complete] A new battle phase begins!**`);
-        return this.start();
-      }
-
-      let winnerMentions = "";
-      if (!this.teams[this.winner].isNpc) {
-        winnerMentions = this.teams[this.winner].userIds
-          .map((userId) => `<@${userId}>`)
-          .join(" ");
-      }
-      this.addToLog(`**Team ${this.winner} has won! ${winnerMentions}**`);
-    } else {
-      this.addToLog("**The battle has ended in a draw!**");
-    }
-    this.ended = true;
-
-    // if winner is npc, no rewards
-    if (this.teams[this.winner] && this.teams[this.winner].isNpc) {
-      this.addToLog("No rewards were given because the winner is an NPC.");
-      return;
-    }
-
-    this.moneyReward = Math.floor(this.baseMoney * this.moneyMultiplier);
-    this.expReward = Math.floor(this.baseExp * this.expMultiplier);
-    // calculate pokemon exp by summing defeated pokemon's levels
-    this.pokemonExpReward = Math.floor(
-      Object.values(this.allPokemon).reduce((acc, pokemon) => {
-        if (pokemon.isFainted) {
-          return acc + (this.minLevel || pokemon.level);
-        }
-        return acc;
-      }, 0) * this.pokemonExpMultiplier
-    );
-
-    this.addToLog(
-      `Winners recieved ${formatMoney(this.moneyReward)}, ${
-        this.expReward
-      } exp, and ${
-        this.pokemonExpReward
-      } BASE Pokemon exp. Losers recieved half the amount.`
-    );
-  }
-
-  createWeather(weatherId, source) {
-    // calculate turns = 10 + number of non-fainted Pokemon
-    const duration =
-      10 +
-      Object.values(this.allPokemon).reduce((acc, pokemon) => {
-        if (!pokemon.isFainted) {
-          return acc + 1;
-        }
-        return acc;
-      }, 0);
-
-    // if weather exists and is same, refresh duration if possible
-    if (this.weather.weatherId === weatherId) {
-      if (this.weather.duration < duration) {
-        this.weather.duration = duration;
-        this.weather.source = source;
-        this.addToLog(`The weather intensified!`);
-        return true;
-      }
-      return false;
-    }
-
-    // apply weather
-    this.weather = {
-      weatherId,
-      duration,
-      source,
-    };
-    switch (weatherId) {
-      case weatherConditions.SUN:
-        this.addToLog(`The sunlight turned harsh!`);
-        break;
-      case weatherConditions.RAIN:
-        this.addToLog(`It started to rain!`);
-        break;
-      case weatherConditions.SANDSTORM:
-        this.addToLog(`A sandstorm kicked up!`);
-        break;
-      case weatherConditions.HAIL:
-        this.addToLog(`It started to hail!`);
-        break;
-      default:
-        break;
-    }
-
-    return true;
-  }
-
-  clearWeather() {
-    if (this.weather.weatherId) {
-      this.addToLog(`The weather cleared up!`);
-    }
-    this.weather = {
-      weatherId: null,
-      duration: 0,
-      source: null,
-    };
-  }
-
-  tickWeather() {
-    if (!this.weather.weatherId) {
-      return;
-    }
-
-    if (!this.isWeatherNegated()) {
-      switch (this.weather.weatherId) {
-        case weatherConditions.SANDSTORM:
-          // tick weather for active Pokemon's team
-          this.addToLog(`The sandstorm rages!`);
-          for (const pokemon of this.parties[this.activePokemon.teamName]
-            .pokemons) {
-            // if pokemon not targetable, skip
-            if (this.isPokemonTargetable(pokemon) === false) {
-              continue;
-            }
-            // if pokemon not rock, steel, or ground, damage 1/16 of max hp
-            if (
-              pokemon.type1 !== pokemonTypes.ROCK &&
-              pokemon.type1 !== pokemonTypes.STEEL &&
-              pokemon.type1 !== pokemonTypes.GROUND &&
-              pokemon.type2 !== pokemonTypes.ROCK &&
-              pokemon.type2 !== pokemonTypes.STEEL &&
-              pokemon.type2 !== pokemonTypes.GROUND
-            ) {
-              pokemon.takeDamage(
-                Math.floor(pokemon.maxHp / 16),
-                this.weather.source,
-                {
-                  type: "weather",
-                }
-              );
-            }
-          }
-          break;
-        case weatherConditions.HAIL:
-          // tick weather for active Pokemon's team
-          this.addToLog(`The hail continues!`);
-          for (const pokemon of this.parties[this.activePokemon.teamName]
-            .pokemons) {
-            // if pokemon not targetable, skip
-            if (this.isPokemonTargetable(pokemon) === false) {
-              continue;
-            }
-            // if pokemon not ice, damage 1/16 of max hp
-            if (
-              pokemon.type1 !== pokemonTypes.ICE &&
-              pokemon.type2 !== pokemonTypes.ICE
-            ) {
-              pokemon.takeDamage(
-                Math.floor(pokemon.maxHp / 16),
-                this.weather.source,
-                {
-                  type: "weather",
-                }
-              );
-            }
-          }
-          break;
-        default:
-          break;
-      }
-    }
-
-    this.weather.duration -= 1;
-    if (this.weather.duration <= 0) {
-      this.clearWeather();
-    }
-  }
-
-  isWeatherNegated() {
-    // for all non-fainted pokemon, check if they have cloud nine or air lock
-    for (const pokemon of Object.values(this.allPokemon)) {
-      if (pokemon.isFainted) {
-        continue;
-      }
-
-      if (
-        pokemon.ability.abilityId === "13" ||
-        pokemon.ability.abilityId === "76"
-      ) {
-        return true;
-      }
-    }
-
-    return false;
-  }
-
-  isNpc(userId) {
-    const user = this.users[userId];
-    return user.npc !== undefined;
-  }
-
-  getEligibleTargets(source, moveId) {
-    const moveData = getMove(moveId);
-    const eligibleTargets = [];
-    const eventArgs = {
-      user: source,
-      moveId,
-      eligibleTargets,
-      shouldReturn: false,
-    };
-    this.eventHandler.emit(battleEventEnum.GET_ELIGIBLE_TARGETS, eventArgs);
-    if (eventArgs.shouldReturn) {
-      return eligibleTargets;
-    }
-
-    let targetParty = null;
-    // use target type to get party
-    switch (moveData.targetType) {
-      case targetTypes.ALLY:
-        targetParty = this.parties[source.teamName];
-        break;
-      case targetTypes.ENEMY:
-        for (const teamName in this.parties) {
-          if (teamName !== source.teamName) {
-            targetParty = this.parties[teamName];
-            break;
-          }
-        }
-        break;
-      default:
-        // return all pokemon
-        for (const teamName in this.parties) {
-          const party = this.parties[teamName];
-          for (const pokemon of party.pokemons) {
-            if (this.isPokemonTargetable(pokemon, moveId)) {
-              eligibleTargets.push(pokemon);
-            }
-          }
-        }
-        return eligibleTargets;
-    }
-
-    // use target position to get valid targets
-    let index;
-    switch (moveData.targetPosition) {
-      case targetPositions.SELF:
-        eligibleTargets.push(source);
-        break;
-      case targetPositions.NON_SELF:
-        for (const pokemon of targetParty.pokemons) {
-          if (this.isPokemonTargetable(pokemon, moveId) && pokemon !== source) {
-            eligibleTargets.push(pokemon);
-          }
-        }
-        break;
-      case targetPositions.FRONT:
-        // break up party into rows
-        // if pokemons exist in row, add to eligible targets
-        // if all pokemon in front row fainted or nonexistent, move to next row
-
-        index = 0;
-        for (let i = 0; i < targetParty.rows; i += 1) {
-          let pokemonFound = false;
-          for (let j = 0; j < targetParty.cols; j += 1) {
-            const pokemon = targetParty.pokemons[index];
-            if (this.isPokemonTargetable(pokemon, moveId)) {
-              pokemonFound = true;
-              eligibleTargets.push(pokemon);
-            }
-            index += 1;
-          }
-          if (pokemonFound) {
-            break;
-          }
-        }
-        break;
-      case targetPositions.BACK:
-        // break up party into rows
-        // if pokemons exist in back row, add to eligible targets
-        // if all pokemon in back row fainted or nonexistent, move to next row
-        index = targetParty.pokemons.length - 1;
-        for (let i = targetParty.rows - 1; i >= 0; i -= 1) {
-          let pokemonFound = false;
-          for (let j = targetParty.cols - 1; j >= 0; j -= 1) {
-            const pokemon = targetParty.pokemons[index];
-            if (this.isPokemonTargetable(pokemon, moveId)) {
-              pokemonFound = true;
-              eligibleTargets.push(pokemon);
-            }
-            index -= 1;
-          }
-          if (pokemonFound) {
-            break;
-          }
-        }
-        break;
-      default:
-        // return all pokemon in party
-        for (const pokemon of targetParty.pokemons) {
-          if (this.isPokemonTargetable(pokemon, moveId)) {
-            eligibleTargets.push(pokemon);
-          }
-        }
-        break;
-    }
-
-    return eligibleTargets;
-  }
-
-  // eslint-disable-next-line class-methods-use-this
-  isPokemonTargetable(pokemon, moveId = null) {
-    if (!pokemon || pokemon.isFainted) {
-      return false;
-    }
-
-    if (!pokemon.targetable) {
-      // special cases
-      // TODO: should this be an event listener?
-
-      // eq + dig
-      if (moveId === "m89" && pokemon.effectIds.burrowed !== undefined) {
-        return true;
-      }
-
-      // bounce + thunder or gust or smackdown
-      if (
-        (moveId === "m87" ||
-          moveId === "m87-1" ||
-          moveId === "m16" ||
-          moveId === "m479") &&
-        pokemon.effectIds.sprungUp !== undefined
-      ) {
-        return true;
-      }
-
-      return false;
-    }
-
-    return true;
-  }
-
-  // eslint-disable-next-line class-methods-use-this
-  isPokemonHittable(pokemon, moveId = null) {
-    if (!pokemon || pokemon.isFainted) {
-      return false;
-    }
-
-    if (!pokemon.hittable) {
-      // special cases
-      // TODO: should this be an event listener?
-
-      // eq + dig
-      if (moveId === "m89" && pokemon.effectIds.burrowed !== undefined) {
-        return true;
-      }
-
-      // bounce + thunder or gust or smackdown
-      if (
-        (moveId === "m87" ||
-          moveId === "m87-1" ||
-          moveId === "m16" ||
-          moveId === "m479") &&
-        pokemon.effectIds.sprungUp !== undefined
-      ) {
-        return true;
-      }
-
-      return false;
-    }
-
-    return true;
-  }
-
-  /**
-   * @template {BattleEventEnum} K
-   * @param {object} param0
-   * @param {K} param0.eventName
-   * @param { BattleEventListenerCallback<K> } param0.callback
-   * @param { BattleEventListenerConditionCallback<K>= } param0.conditionCallback function that returns true if the event should be executed. If undefined, always execute for event.
-   * @returns {string} listenerId
-   */
-  registerListenerFunction({ eventName, callback, conditionCallback }) {
-    return this.eventHandler.registerListener(eventName, {
-      isNewListener: true,
-      execute: callback,
-      conditionCallback,
-    });
-  }
-
-  /**
-   * @param {string?=} listenerId
-   */
-  unregisterListener(listenerId) {
-    if (!listenerId) {
-      return;
-    }
-    this.eventHandler.unregisterListener(listenerId);
-  }
-
-  /**
-   * @template {BattleEventEnum} K
-   * @param {K} eventName
-   * @param {BattleEventArgsWithoutEventName<K>} args
-   */
-  emitEvent(eventName, args) {
-    this.eventHandler.emit(eventName, args || {});
-  }
-
-  clearLog() {
-    this.log = [];
-  }
-
-  addToLog(message) {
-    this.log.push(message);
   }
 }
 
@@ -3687,8 +1132,8 @@ const onBattleTowerAccept = async ({ stateId = null, user = null } = {}) => {
  *
  * @param {object} param0
  * @param {string?=} param0.stateId
- * @param {User?=} param0.user
- * @returns
+ * @param {any?=} param0.user
+ * @returns {Promise<any>}
  */
 const buildBattleTowerSend = async ({ stateId = null, user = null } = {}) => {
   // get state
@@ -3751,8 +1196,6 @@ const buildBattleTowerSend = async ({ stateId = null, user = null } = {}) => {
 };
 
 module.exports = {
-  Battle,
-  Pokemon,
   getStartTurnSend,
   buildPveSend,
   buildDungeonSend,
